code re-org have slightly less pub(crate) items
