+++ src/lib.rs
@@ -51,7 +51,6 @@ mod macros;
 mod conversion;
 mod function;
 mod lua;
-mod methods;
 mod multi;
 mod scope;
 mod string;
@@ -68,14 +67,13 @@ mod tests;
 pub use error::{Error, ExternalError, ExternalResult, Result};
 pub use function::Function;
 pub use lua::Lua;
-pub use methods::{MetaMethod, UserDataMethods};
 pub use multi::Variadic;
 pub use scope::Scope;
 pub use string::String;
 pub use table::{Table, TablePairs, TableSequence};
 pub use thread::{Thread, ThreadStatus};
 pub use types::{Integer, LightUserData, Number, RegistryKey};
-pub use userdata::{AnyUserData, UserData};
+pub use userdata::{AnyUserData, MetaMethod, UserData, UserDataMethods};
 pub use value::{FromLua, FromLuaMulti, MultiValue, Nil, ToLua, ToLuaMulti, Value};
 
 pub mod prelude;
